// Qanat.Web/src/app/shared/generated/model/external-map-layer-simple-dto.ts
export class ExternalMapLayerSimpleDto { 
    ExternalMapLayerID?: number;
    GeographyID?: number;
    ExternalMapLayerDisplayName?: string | null;
    ExternalMapLayerTypeID?: number;
    ExternalMapLayerTypeDisplayName?: string | null;
    ExternalMapLayerURL?: string | null;
    LayerIsOnByDefault?: boolean;
    IsActive?: boolean;
    ExternalMapLayerDescription?: string | null;
    PopUpField?: string | null;
    MinZoom?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}